Fix : exec-missing, wait-killed
